If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,16 +1,19 @@
+#include <stdlib.h>
 #include "function_pointers.h"
 
 /**
-  * array_iterator - ...
-  * @array: ...
-  * @size: ...
-  * @action: ...
-  *
-  * Return: ...
-  */
+ * array_iterator - a function that executes a function given as
+ * a parameter on each element of an array
+ *
+ * @array: array of integers
+ * @size: size of the array
+ * @action: pointer to a function
+ *
+ * Return: (void)
+ */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	if (array != NULL && action != NULL && size > 0)
 	{
